refactor: remove integration build tags and fix linter issues
- Remove //go:build integration tags from all integration test files
- Update CI workflow to remove -tags=integration flag
- Create shared helpers_test.go for common integration test functions
- Remove duplicate setupFreshDatabase functions
- Fix import aliases to use 'qt' for quicktest
- Remove unused createTenantViaCLI function
- Move registry registration from init() to test function
- Fix import formatting with gci
- Remove unused imports

Integration tests now skip gracefully when required environment
variables (POSTGRES_TEST_DSN) are not set, eliminating the need
for build tags while maintaining proper test isolation.
